+++ modules/tinymce/src/plugins/advlist/test/ts/browser/NoneditableRootTest.ts
@@ -1,7 +1,7 @@
 import { UiFinder } from '@ephox/agar';
 import { context, describe, it } from '@ephox/bedrock-client';
 import { SugarBody } from '@ephox/sugar';
-import { TinyAssertions, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-mcagar';
+import { TinyAssertions, TinyHooks, TinySelections, TinyState, TinyUiActions } from '@ephox/wrap-mcagar';
 
 import Editor from 'tinymce/core/api/Editor';
 import Plugin from 'tinymce/plugins/lists/Plugin';
@@ -15,12 +15,6 @@ describe('browser.tinymce.plugins.advlist.NoneditableRootTest', () => {
     base_url: '/project/tinymce/js/tinymce'
   }, [ Plugin ], true);
 
-  const withNoneditableRootEditor = (editor: Editor, f: (editor: Editor) => void) => {
-    editor.getBody().contentEditable = 'false';
-    f(editor);
-    editor.getBody().contentEditable = 'true';
-  };
-
   context('List ui controls', () => {
     const initialListContent = '<ol><li>a</li></ol>';
     const setupEditor = (editor: Editor) => {
@@ -29,7 +23,7 @@ describe('browser.tinymce.plugins.advlist.NoneditableRootTest', () => {
     };
 
     it('TINY-9458: List buttons numlist/bullist should be disabled', () => {
-      withNoneditableRootEditor(hook.editor(), (editor) => {
+      TinyState.withNoneditableRootEditor<Editor>(hook.editor(), (editor) => {
         setupEditor(editor);
 
         UiFinder.exists(SugarBody.body(), 'div[title="Numbered list"][aria-disabled="true"]');
@@ -38,7 +32,7 @@ describe('browser.tinymce.plugins.advlist.NoneditableRootTest', () => {
     });
 
     it('TINY-9458: Outdent/indent buttons should be noop', () => {
-      withNoneditableRootEditor(hook.editor(), (editor) => {
+      TinyState.withNoneditableRootEditor<Editor>(hook.editor(), (editor) => {
         setupEditor(editor);
 
         TinyUiActions.clickOnToolbar(editor, 'button[title="Decrease indent"]');
@@ -50,14 +44,12 @@ describe('browser.tinymce.plugins.advlist.NoneditableRootTest', () => {
     });
 
     it('TINY-9458: Context menu list properties should be disabled', async () => {
-      const editor = hook.editor();
+      await TinyState.withNoneditableRootEditorAsync<Editor>(hook.editor(), async (editor) => {
+        setupEditor(editor);
 
-      setupEditor(editor);
-
-      editor.getBody().contentEditable = 'false';
-      await TinyUiActions.pTriggerContextMenu(editor, 'li', '.tox-silver-sink [role="menuitem"]:contains("List properties...")');
-      UiFinder.exists(SugarBody.body(), '[role="menuitem"][aria-disabled="true"]:contains("List properties...")');
-      editor.getBody().contentEditable = 'true';
+        await TinyUiActions.pTriggerContextMenu(editor, 'li', '.tox-silver-sink [role="menuitem"]:contains("List properties...")');
+        UiFinder.exists(SugarBody.body(), '[role="menuitem"][aria-disabled="true"]:contains("List properties...")');
+      });
     });
   });
 });