+++ modules/tinymce/src/plugins/directionality/test/ts/browser/NoneditableRootTest.ts
@@ -1,0 +1,36 @@
+import { UiFinder } from '@ephox/agar';
+import { describe, it } from '@ephox/bedrock-client';
+import { SugarBody } from '@ephox/sugar';
+import { TinyHooks, TinySelections, TinyState } from '@ephox/wrap-mcagar';
+
+import Editor from 'tinymce/core/api/Editor';
+import Plugin from 'tinymce/plugins/directionality/Plugin';
+
+describe('browser.tinymce.plugins.directionality.NoneditableRootTest', () => {
+  const hook = TinyHooks.bddSetup<Editor>({
+    plugins: 'directionality',
+    toolbar: 'ltr rtl',
+    base_url: '/project/tinymce/js/tinymce'
+  }, [ Plugin ], true);
+
+  it('TINY-9669: Disable ltr button on noneditable content', () => {
+    TinyState.withNoneditableRootEditor(hook.editor(), (editor) => {
+      editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+      TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Left to right"][aria-disabled="true"]');
+      TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Left to right"][aria-disabled="false"]');
+    });
+  });
+
+  it('TINY-9669: Disable rtl button on noneditable content', () => {
+    TinyState.withNoneditableRootEditor(hook.editor(), (editor) => {
+      editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+      TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Right to left"][aria-disabled="true"]');
+      TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+      UiFinder.exists(SugarBody.body(), '[aria-label="Right to left"][aria-disabled="false"]');
+    });
+  });
+});
+