+++ modules/tinymce/src/plugins/emoticons/main/ts/ui/Buttons.ts
@@ -1,4 +1,18 @@
 import Editor from 'tinymce/core/api/Editor';
+import { Menu, Toolbar } from 'tinymce/core/api/ui/Ui';
+
+const onSetupEditable = (editor: Editor) => (api: Toolbar.ToolbarButtonInstanceApi | Menu.MenuItemInstanceApi): VoidFunction => {
+  const nodeChanged = () => {
+    api.setEnabled(editor.selection.isEditable());
+  };
+
+  editor.on('NodeChange', nodeChanged);
+  nodeChanged();
+
+  return () => {
+    editor.off('NodeChange', nodeChanged);
+  };
+};
 
 const register = (editor: Editor): void => {
   const onAction = () => editor.execCommand('mceEmoticons');
@@ -6,13 +20,15 @@ const register = (editor: Editor): void => {
   editor.ui.registry.addButton('emoticons', {
     tooltip: 'Emojis',
     icon: 'emoji',
-    onAction
+    onAction,
+    onSetup: onSetupEditable(editor)
   });
 
   editor.ui.registry.addMenuItem('emoticons', {
     text: 'Emojis...',
     icon: 'emoji',
-    onAction
+    onAction,
+    onSetup: onSetupEditable(editor)
   });
 };