+++ modules/tinymce/src/plugins/table/test/ts/browser/ui/TableRowDialogTest.ts
@@ -2,7 +2,7 @@ import { UiFinder } from '@ephox/agar';
 import { afterEach, describe, it } from '@ephox/bedrock-client';
 import { Arr } from '@ephox/katamari';
 import { SugarBody, SugarElement, SugarNode } from '@ephox/sugar';
-import { TinyAssertions, TinyHooks, TinySelections } from '@ephox/wrap-mcagar';
+import { TinyAssertions, TinyHooks, TinySelections, TinyState } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -405,7 +405,7 @@ describe('browser.tinymce.plugins.table.TableRowDialogTest', () => {
   });
 
   it('TINY-9459: Should not open table row properties dialog on noneditable root', () => {
-    TableTestUtils.withNoneditableRootEditor(hook.editor(), (editor) => {
+    TinyState.withNoneditableRootEditor(hook.editor(), (editor) => {
       editor.setContent('<table><tbody><tr><td>x</td></tr></tbody></table>');
       TinySelections.setCursor(editor, [ 0, 0, 0, 0, 0 ], 0);
       editor.execCommand('mceTableRowProps');