+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/core/ChoiceControlsTest.ts
@@ -1,9 +1,9 @@
-import { UiFinder, Waiter } from '@ephox/agar';
+import { Keys, UiFinder, Waiter } from '@ephox/agar';
 import { beforeEach, context, describe, it } from '@ephox/bedrock-client';
 import { Arr, Optional } from '@ephox/katamari';
 import { PlatformDetection } from '@ephox/sand';
-import { Attribute } from '@ephox/sugar';
-import { McEditor, TinyAssertions, TinyHooks, TinySelections, TinyUiActions } from '@ephox/wrap-mcagar';
+import { Attribute, SugarBody } from '@ephox/sugar';
+import { McEditor, TinyAssertions, TinyHooks, TinySelections, TinyState, TinyUiActions } from '@ephox/wrap-mcagar';
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
@@ -143,6 +143,30 @@ describe('browser.tinymce.themes.silver.editor.core.ChoiceControlsTest', () => {
 
         await pAssertOptions(editor, menuSpec.menuSelector, [ '1', '1.1', '1.2', '1.3', '1.4', '1.5', '2' ], Optional.some('1.4'));
         menuSpec.close(editor, 'Line height');
+      });
+
+      it('TINY-9669: Disable lineheight button on noneditable content', () => {
+        TinyState.withNoneditableRootEditor<Editor>(hook.editor(), (editor) => {
+          editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+          TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+          UiFinder.exists(SugarBody.body(), '[aria-label="Line height"]:disabled');
+          TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+          UiFinder.exists(SugarBody.body(), '[aria-label="Line height"]:not(:disabled)');
+        });
+      });
+
+      it('TINY-9669: Disable lineheight menuitem on noneditable content', async () => {
+        await TinyState.withNoneditableRootEditorAsync<Editor>(hook.editor(), async (editor) => {
+          editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+          TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+          TinyUiActions.clickOnMenu(editor, 'button:contains("Format")');
+          await TinyUiActions.pWaitForUi(editor, '[role="menu"] [title="Line height"][aria-disabled="true"]');
+          TinyUiActions.keystroke(editor, Keys.escape());
+          TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+          TinyUiActions.clickOnMenu(editor, 'button:contains("Format")');
+          await TinyUiActions.pWaitForUi(editor, '[role="menu"] [title="Line height"][aria-disabled="false"]');
+          TinyUiActions.keystroke(editor, Keys.escape());
+        });
       });
     });
 
@@ -285,6 +309,32 @@ describe('browser.tinymce.themes.silver.editor.core.ChoiceControlsTest', () => {
 
         editor.formatter.apply('lang', { value: 'zh' });
         await TinyUiActions.pWaitForUi(editor, '.tox-tbtn.tox-tbtn--enabled');
+      });
+
+      context('Noneditable', () => {
+        it('TINY-9669: Disable language button on noneditable content', () => {
+          TinyState.withNoneditableRootEditor<Editor>(hook.editor(), (editor) => {
+            editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+            TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+            UiFinder.exists(SugarBody.body(), '[aria-label="Language"]:disabled');
+            TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+            UiFinder.exists(SugarBody.body(), '[aria-label="Language"]:not(:disabled)');
+          });
+        });
+
+        it('TINY-9669: Disable language menuitem on noneditable content', async () => {
+          await TinyState.withNoneditableRootEditorAsync<Editor>(hook.editor(), async (editor) => {
+            editor.setContent('<div>Noneditable content</div><div contenteditable="true">Editable content</div>');
+            TinySelections.setSelection(editor, [ 0, 0 ], 0, [ 0, 0 ], 2);
+            TinyUiActions.clickOnMenu(editor, 'button:contains("Format")');
+            await TinyUiActions.pWaitForUi(editor, '[role="menu"] [title="Language"][aria-disabled="true"]');
+            TinyUiActions.keystroke(editor, Keys.escape());
+            TinySelections.setSelection(editor, [ 1, 0 ], 0, [ 1, 0 ], 2);
+            TinyUiActions.clickOnMenu(editor, 'button:contains("Format")');
+            await TinyUiActions.pWaitForUi(editor, '[role="menu"] [title="Language"][aria-disabled="false"]');
+            TinyUiActions.keystroke(editor, Keys.escape());
+          });
+        });
       });
     });